Adds ability for players to have multiple goals
`Referees` now have a `Config` struct to store a configuration of how
the game should be run.
- `Config` stores any data the referee might need in only some
  configurations
This also refactors the default game to no longer use a Hashset to store
the players who have reached their first goal. Instead opting to store
the number of goals they have reached.